Handle scenario where unique constraint with condition is provided
Postgres does not support unique constraints with conditions, however,
Django allows it to be possible to specify a unique constraint with a
condition. In postgres, these are represented as unique indexes with
conditions instead

Here, we modify the operations to add or remove constraints to align
with django operations behaviour - adding or removing the unique index -
however with "if exists / not exists" and concurrently added to
these.